Written assignment: How does the world wide web work?
- What is a URL?
URL stands for a Universal or Uniform Resource Locator. It helps any computer to locate information stored on any other computer by having a particular page address. URL gives information to computer system such as where to find a page on user’s computer, how to access it, and what to do with it to display it correctly.  

- What is HTTP?
HTTP is an abbreviation for Hyper Text Transfer Protocol, which is a way of formatting and transmitting information through a simple “conversation” between two different computers: a client running a program called a web browser, and a server or web server. HTTP is the simple way for one computer to ask another computer for Web pages.

- What is a browser?
A browser, also known as a web browser, is computer software that people use to search for information on the Internet. Basically, web browsers and web servers converse each other in HTTP, with the browser sending requests for particular information needed and the server sending them if it can find the requested information. While the browser is sending requests, it explains what software it is, what operating system it’s running, which character sets such as foreign fonts it can accept, which forms of compressed file it can accept in gzip and deflate, and which file it wants. On the other hand, the server sends a compressed file in gzip by explaining how long the data is and what format it is in.

- What’s the difference between the web and the Internet?
The Internet is the worldwide computer network linked mostly by telephone lines whereas the Web is the worldwide computer system, which links text pages, digital photographs, music files, videos, and animations people can access over the Internet. Furthermore, the Internet allows one computer user to connect with another computer user all over the world such as an email. And, the Web allows the user to use a set of data and information through a collection of Web pages on the Internet. In addition to the Web, a website is a publicly accessible space in which users can read information that other users has stored.